+++ AboutUs.html
@@ -20,7 +20,6 @@
         <header id="header-wrapper" class="bg-white"></header>
         <!--/header-->
 
-        <main>
         <!--Portada-->
         <section id="portada">
             <div id="wave-portada">
@@ -192,7 +191,7 @@
             </ul>
         </section>
         <!--/cards-->
-    </main>
+
         <!--footer-->
         <footer class="bg-white" id="footer-1"></footer>
         <!--footer-->
@@ -207,8 +206,5 @@
         <script>
             AOS.init();
         </script>
-        <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
     </body>
 </html>
