+++ SearchPage.html
@@ -19,12 +19,8 @@
         <!--header-->
         <header id="header-wrapper" class="bg-white"></header>
         <!--/header-->
-        <main>
-
-        <!-- Cards -->
-        <div id="cardCreate" class="container">
-            <!--search-container-->
 <section class="search-container-inner">
+    <!--search-container-->
     <!--*Aqui empieza EL BUSCADOR XL -->
     <div class="search-container">
         <!--Div de buscar por especialidad-->
@@ -83,11 +79,10 @@
 
 <section>
 
-        </div>
+        <!-- Cards -->
+        <div id="cardCreate" class="container"></div>
         <!-- Cards -->
 </section>
-
-</main>
         <!--footer-->
         <footer class="bg-white" id="footer-1"></footer>
         <!--footer-->
@@ -95,9 +90,7 @@
         <!--Menu hamburguesa-->
         <div id="hamburger-menu" class="main_hamburger_menu_container"></div>
         <!--/Menu hamburguesa-->
-        <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
+
         <script src="./assets/js/main.js" type="module"></script>
     </body>
 </html>
