+++ SignInPatient.html
@@ -25,7 +25,7 @@
           <div class="container-form">
               <div class="column-wrapper">
                   <div class="cont-izq">
-                   <img id="form-imagen" src="./assets/img/img-form.jpg" alt="">
+                   <img id="form-imagen" src="./assets/img/img-form.png" alt="">
                   </div>
                   <div class="cont-der">
                     <div id="wrapper-form">
@@ -118,9 +118,6 @@
       <!--Menu hamburguesa-->
       <div id="hamburger-menu" class="main_hamburger_menu_container"></div>
       <!--/Menu hamburguesa-->
-      <script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-      <script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-      <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
       <script src="./assets/js/SignInPatient.js"></script>
   <script src="./assets/js/main.js" type="module"></script>
 </body>
