+++ assets/css/SearchPage.css
@@ -12,6 +12,7 @@ hr {
 
 .search-container-inner{
     margin-top: 100px;
+    margin-left: 5rem;
 }
