--- assets/css/profileDoctor.css
@@ -1,431 +1,0 @@
-.Portada-profile {
-    width: 100%;
-    height: 3000px;
-}
-
-.portada-container { 
-    width: 100%;
-    height: 100%;
-}
-
-.portada-wrapper {
-    width: 100%;
-    height: 100%;
-    display: flex;
-    flex-direction: row;
-}
-
-.port-cont1 {
-    width: 35%;
-    height: 100%;
-}
-
-
-.div-profile-container {
-    position: relative;
-    position: -webkit-sticky;
-    position: sticky;
-    top: 103px;
-}
-
-
-.img-profile-wrapper {
-    width: 100%;
-    height: 100%;
-    overflow: hidden;
-}
-
-.img-profile-page {
-    max-height: 610px;
-    width: 100%;
-    object-fit: cover;
-}
-
-.profile-banner {
-    display: block;
-    max-width: 104%;
-    position: absolute;
-    bottom: 0;
-    max-height: 198px;
-}
-
-.doctor-name-profile-wrapper {
-    width: 100%;
-    bottom: 0;
-    text-align: center;
-    position: absolute;
-    padding-bottom: 25px;
-}
-
-.name-doctor-profile-card {
-    font-weight: 800;
-    color: #FFF;
-    font-family: Lexend;
-}
-
-.especialidad-profile-card {
-    color: #fff;
-}
-
-.star-rate {
-    font-family: lexend;
-    font-weight: 300;
-    color: #fff;
-}
-
-.star-img {
-    margin-left: 12px;
-}
-
-
-.especialidad-star-profile-wrapper {
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-content: center;
-    align-items: center;
-}
-
-.rating-star-profile {
-    display: flex;
-    flex-direction: row;
-}
-
-
-.port-cont2 {
-    width: 65%;
-    height: 100%;
-}
-
-.perfil-container-profile {
-    margin: 30px 80px 80px 80px;
-}
-
-
-.sub-logo-especialidad {
-    width: 120px;
-    position: absolute;
-    right: 24%;
-    top: 24%;
-}
-
-.main-container-profile {
-    margin: 2rem;
-}
-
-.subtitle-profile {
-    font-size: 35px;
-    font-weight: 700;
-    font-family: lexend;
-    color: #79DAE8;
-    padding-top: 30px;
-}
-
-.name-doctor-profile {
-    font-size: 30px;
-    font-weight: 700;
-    font-family: lexend;
-}
-
-.especialidad-description {
-    font-size: 30px;
-    font-weight: 300;
-    font-family: lexend;
-}
-
-.especialidad-description-uni {
-    font-size: 26px;
-    font-weight: 300;
-    font-family: lexend;
-}
-
-.especialidad-section {
-    background-color: #E8F9FD;
-    padding-bottom: 30px;
-    padding-left: 2rem;
-    padding-right: 2rem;
-}
-
-.especialidad-container {
-    margin: 30px 80px 30px 80px;
-    height: 100%;
-}
-
-.enfermedades-4-container{
-    margin-bottom: 30px;
-}
-
-
-.enfermedades-4{
-    display: flex;
-    flex-direction: row;
-    height: 100%;
-}
-
-.enfermedades1-2{
-    width: 50%;
-    height: 100%;
-    text-align: center;
-}
-
-.enfermedades3-4{
-    width: 50%;
-    height: 100%;
-    text-align: center;
-}
-
-.items-enfermedades-wrapper{
-    display: flex;
-    height: 100%;
-}
-
-.items-enfermedades{
-    width: 270px;
-    margin-top: 30px;
-    padding-top: 10px;
-    padding-bottom: 10px;
-    border-radius: 50px;
-    border-style: solid;
-    border: solid #79DAE8;
-    align-content: center;
-    font-family: lexend;
-    font-size: 26px;
-    font-weight: 400;
-    -webkit-border-radius: 50px;
-    -moz-border-radius: 50px;
-    -ms-border-radius: 50px;
-    -o-border-radius: 50px;
-}
-
-.section-biografia{
-    padding-bottom: 30px;
-    padding-left: 2rem;
-    padding-right: 2rem;
-}
-
-.biografia-container{
-    margin: 30px 80px 30px 80px;
-    height: 100%;
-}
-
-.biografia-text{
-    font-size: 26px;
-    font-family: lexend;
-    font-weight: 300;
-    text-align: justify 12;
-}
-
-.section-galeria {
-    background-color: #E8F9FD;
-    padding-bottom: 30px;
-    padding-left: 2rem;
-    padding-right: 2rem;
-}
-
-.galeria-container {
-    margin: 30px 80px 30px 80px;
-    height: 100%;
-}
-
-/* s
-s
-s */
-
-img {
-	max-width: 100%;
-}
-
-
-/*
- sprite svg
- */
- .sprite-svg {
-    display: none;
-  }
-  /*
-   title
-   */
-  h1 {
-    padding: 40px 0 40px;
-    font-size: 40px;
-    text-align: center;
-    text-transform: uppercase;
-  }
-  /*
-   images wrapper
-   */
-  .wrapper {
-    display: block;
-    width: 100%;
-    max-width: 1200px;
-    columns: 1;
-    column-gap: 0;
-    break-inside: avoid;
-    margin: 20px auto 50px;
-  }
-  @media (min-width: 600px) {
-    .wrapper {
-      columns: 2;
-    }
-  }
-  @media (min-width: 800px) {
-    .wrapper {
-      columns: 2;
-    }
-  }
-  .grid-image {
-    float: left;
-    width: 100%;
-    padding: 4px;
-  }
-  .grid-image img {
-    float: left;
-    width: 100%;
-  }
-  .zoom-wrapper {
-    z-index: 200;
-    display: none;
-    position: fixed;
-    left: 0;
-    top: 0;
-    bottom: 0;
-    width: 100%;
-    overflow: auto;
-    background-color: black;
-    background-color: rgba(0, 0, 0, 0.6);
-  }
-  .zoom-wrapper:target {
-    display: flex;
-  }
-  @keyframes top-animation {
-    from {
-      top: -100px;
-      opacity: 0;
-    }
-    to {
-      top: 0px;
-      opacity: 1;
-    }
-  }
-  .zoom-wrapper__image {
-    position: relative;
-    width: 80%;
-    max-width: 1000px;
-    height: 100%;
-    margin: auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    /* animation */
-    animation-name: top-animation;
-    animation-duration: 0.5s;
-  }
-  .cell {
-    box-flex: 1;
-    display: flex;
-    flex: 1 1 auto;
-    justify-content: center;
-  }
-  .zoom-image {
-    max-height: 100%;
-    margin: auto;
-    transform: scale(60%);
-    -webkit-transform: scale(60%);
-    -moz-transform: scale(60%);
-    -ms-transform: scale(60%);
-    -o-transform: scale(60%);
-}
-  /*
-   arrow navigation
-   */
-  .navigation-arrow {
-    position: fixed;
-    top: 50%;
-    transform: translateY(-50%);
-    perspective: 1000px;
-    perspective-origin: 100% 50%;
-  }
-  .navigation-arrow__left {
-    left: 0;
-  }
-  .navigation-arrow__next {
-    right: 0;
-  }
-  .navigation-arrow__image {
-    position: absolute;
-    width: 100px;
-    height: 100px;
-    overflow: hidden;
-    transition: transform 0.3s;
-  }
-  .navigation-arrow__prev .navigation-arrow__image {
-    left: 40px;
-    transform: rotateY(90deg);
-    transform-origin: 0% 50%;
-  }
-  .navigation-arrow__next .navigation-arrow__image {
-    right: 40px;
-    transform: rotateY(-90deg);
-    transform-origin: 100% 50%;
-  }
-  .navigation-arrow__image img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-  .navigation-arrow__prev:hover .navigation-arrow__image, .navigation-arrow__next:hover .navigation-arrow__image {
-    transform: rotateY(0deg);
-  }
-  .navigation-arrow__icon {
-    position: relative;
-    display: block;
-    width: 40px;
-    height: 100px;
-    padding: 6px;
-    background-color: #ffffff;
-  }
-  .navigation-arrow__icon svg.icon {
-    top: 50%;
-    transform: translateY(-50%);
-    fill: #000000;
-    position: relative;
-    stroke-width: 2;
-    stroke: #000000;
-  }
-  .zoom-wrapper__close {
-    position: fixed;
-    top: 10px;
-    right: 10px;
-    z-index: 1;
-    color: #ffffff;
-    font-size: 40px;
-    text-decoration: none;
-    text-shadow: 0px 0px 2px rgba(0, 0, 0, 0.3);
-  }
-  
-
-
-
-/* s */
-/* s */
-/* s */
-
-
-@media (max-width:1200px) {
-
-    .name-doctor-profile-card {
-        font-size: 24px;
-    }
-
-    .especialidad-profile-card .star-img {
-        font-size: 18px;
-    }
-
-    .doctor-name-profile-wrapper {
-        padding-bottom: 5px;
-    }
-
-}
-
-
-@media (max-width:910px) {}
