+++ assets/js/SignInDoc.js
@@ -1,5 +1,3 @@
-import SignInDoctor from "./libs/SignInDoctor.js";
-
 const formulario = document.getElementById('formulario'); //Se define una constante para poder acceder al formulario
 const inputs = document.querySelectorAll('#formulario input'); // Se define una constante para poder acceder a todos los inputs del formulario
 
@@ -23,7 +21,7 @@ const campos = {
 }
 
 const validarFormulario = (e) => {
-	switch (e.target.id) {
+	switch (e.target.name) {
 		
 		case "nombre":
 			validarCampo(expresiones.nombre, e.target, 'nombre');
@@ -91,14 +89,12 @@ formulario.addEventListener('submit', (e) => {
 
 	
 	if(campos.nombre && campos.apellido1 &&  campos.apellido2 && campos.correo && campos.password && terminos.checked){
-		//formulario.reset();
+		formulario.reset();
 		document.getElementById('formulario__mensaje').classList.remove('formulario__mensaje-activo');
 		document.getElementById('formulario__mensaje-exito').classList.add('formulario__mensaje-exito-activo');
 		setTimeout(() => {
 			document.getElementById('formulario__mensaje-exito').classList.remove('formulario__mensaje-exito-activo');
 		}, 4000);
-		
-		SignInDoctor(formulario);
 	} else {
 		document.getElementById('formulario__mensaje').classList.add('formulario__mensaje-activo');
 	}
