+++ assets/js/templates/Header.js
@@ -4,7 +4,7 @@ const Header = function(){
 
     // al innerHTML del nodo creado le inserto el codigo html del componente
     header.innerHTML= `
-    <div class="container-header sticky">
+    <div class="container">
             <nav class="navbar navbar-expand-lg navbar-light bg-white header-space flex-nowrap">
                 <a class="navbar-brand color-azul" href="index.html">
                     <svg width="262" height="45" viewBox="0 0 262 60" fill="none" xmlns="http://www.w3.org/2000/svg">
@@ -27,64 +27,43 @@ const Header = function(){
                     </svg>
                 </a>
                 <div class="w-lg-100">
-                <div class="navbar-collapse justify-content-end" id="navbarNavAltMarkup">
-                    <div class="navbar-nav d-flex align-items-center">
-            
-                        <!--Enlace principal-->
-                        <div class="nav-link d-none d-lg-flex flex-row align-items-center">
-                            <div class="icon-estetoscopio mr-3">
-                                <svg width="70" height="70" viewBox="0 0 70 70" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                    <path d="M56.96 20.229C55.3744 20.229 53.8245 20.6992 52.5062 21.58C51.1879 22.4609 50.1603 23.713 49.5536 25.1778C48.9468 26.6427 48.7881 28.2545 49.0974 29.8096C49.4067 31.3647 50.1703 32.7931 51.2914 33.9143C52.4126 35.0354 53.841 35.7989 55.3961 36.1082C56.9512 36.4175 58.5631 36.2587 60.0279 35.652C61.4927 35.0452 62.7448 34.0176 63.6256 32.6993C64.5065 31.381 64.9766 29.831 64.9766 28.2455C64.9766 26.1194 64.132 24.0803 62.6286 22.577C61.1251 21.0736 59.0861 20.229 56.96 20.229ZM56.9599 31.8867C56.2396 31.8867 55.5354 31.6731 54.9366 31.2729C54.3377 30.8727 53.8709 30.3039 53.5953 29.6385C53.3197 28.973 53.2476 28.2407 53.3881 27.5343C53.5287 26.8278 53.8756 26.1789 54.3849 25.6697C54.8943 25.1604 55.5432 24.8136 56.2497 24.6731C56.9562 24.5327 57.6884 24.6049 58.3539 24.8806C59.0193 25.1563 59.588 25.6231 59.9881 26.2221C60.3882 26.821 60.6017 27.5252 60.6016 28.2455C60.6005 29.2109 60.2164 30.1365 59.5337 30.8192C58.851 31.5018 57.9253 31.8857 56.9599 31.8867Z" fill="#79DAE8" />
-                                    <path d="M52.8194 39.8084V47.3437C52.8152 51.152 51.3006 54.8032 48.6077 57.4962C45.9149 60.1892 42.2638 61.7041 38.4554 61.7085C30.8432 61.7085 24.5133 55.6427 24.1209 48.0879C33.5968 47.0107 40.9468 38.8324 40.9471 29.2479V9.37539C40.9464 8.81397 40.7231 8.27573 40.3262 7.87875C39.9292 7.48176 39.3909 7.25847 38.8295 7.25784H34.1584C33.689 6.40902 32.9508 5.74026 32.0598 5.3567C31.1689 4.97314 30.1758 4.89654 29.2366 5.13892C28.2974 5.38131 27.4654 5.92895 26.8713 6.69574C26.2773 7.46253 25.9549 8.405 25.9548 9.37498C25.9548 10.345 26.2771 11.2875 26.8711 12.0543C27.4651 12.8211 28.2971 13.3688 29.2363 13.6113C30.1755 13.8537 31.1686 13.7772 32.0596 13.3937C32.9505 13.0102 33.6888 12.3414 34.1583 11.4927H36.7127V29.2479C36.7145 31.1847 36.3347 33.1028 35.5948 34.8927C34.8549 36.6826 33.7695 38.3091 32.4007 39.6792C31.0318 41.0494 29.4064 42.1364 27.6172 42.878C25.8281 43.6196 23.9103 44.0013 21.9735 44.0013C20.0368 44.0013 18.119 43.6196 16.3298 42.878C14.5407 42.1364 12.9152 41.0494 11.5464 39.6792C10.1775 38.3091 9.09216 36.6826 8.35228 34.8927C7.61241 33.1028 7.23255 31.1847 7.23441 29.2479V11.4927H9.78881C10.2583 12.3414 10.9966 13.0101 11.8875 13.3936C12.7785 13.7771 13.7716 13.8536 14.7108 13.6111C15.65 13.3686 16.482 12.8209 17.076 12.0541C17.6699 11.2872 17.9922 10.3447 17.9921 9.37472C17.9921 8.40474 17.6697 7.46228 17.0756 6.69551C16.4815 5.92873 15.6495 5.38112 14.7103 5.13877C13.7711 4.89641 12.7779 4.97305 11.887 5.35663C10.9961 5.74022 10.2579 6.409 9.78854 7.25784H5.1174C4.55601 7.25849 4.01779 7.4818 3.62084 7.87878C3.22389 8.27576 3.00061 8.81399 3 9.37539V29.2479C3.00014 38.8559 10.3773 47.0425 19.8825 48.0938C20.2812 57.9792 28.5145 65.9425 38.4554 65.9425C48.7104 65.9425 57.0538 57.5995 57.0538 47.3437V39.8084C55.6557 40.0829 54.2176 40.0829 52.8194 39.8084Z" fill="#79DAE8" />
-                                </svg>
-                            </div>
-            
-                            <div class="dropdown">
-                                <button class="btn_drop" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Soy Doctor</button>
-                                <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-                                    <button class="dropdown-item" data-toggle="modal" data-target="#loginDoctor">Ingresar</button>
-                                    <a class="dropdown-item" href="SignInDoc.html">Registrarse</a>
+                    <div class="navbar-collapse justify-content-end" id="navbarNavAltMarkup">
+                        <div class="navbar-nav d-flex align-items-center">
+                            <a href="SignInDoc.html" class="nav-link d-none d-lg-flex flex-row align-items-center">
+                                <div class="icon-estetoscopio mr-3">
+                                    <svg width="70" height="70" viewBox="0 0 70 70" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                        <path d="M56.96 20.229C55.3744 20.229 53.8245 20.6992 52.5062 21.58C51.1879 22.4609 50.1603 23.713 49.5536 25.1778C48.9468 26.6427 48.7881 28.2545 49.0974 29.8096C49.4067 31.3647 50.1703 32.7931 51.2914 33.9143C52.4126 35.0354 53.841 35.7989 55.3961 36.1082C56.9512 36.4175 58.5631 36.2587 60.0279 35.652C61.4927 35.0452 62.7448 34.0176 63.6256 32.6993C64.5065 31.381 64.9766 29.831 64.9766 28.2455C64.9766 26.1194 64.132 24.0803 62.6286 22.577C61.1251 21.0736 59.0861 20.229 56.96 20.229ZM56.9599 31.8867C56.2396 31.8867 55.5354 31.6731 54.9366 31.2729C54.3377 30.8727 53.8709 30.3039 53.5953 29.6385C53.3197 28.973 53.2476 28.2407 53.3881 27.5343C53.5287 26.8278 53.8756 26.1789 54.3849 25.6697C54.8943 25.1604 55.5432 24.8136 56.2497 24.6731C56.9562 24.5327 57.6884 24.6049 58.3539 24.8806C59.0193 25.1563 59.588 25.6231 59.9881 26.2221C60.3882 26.821 60.6017 27.5252 60.6016 28.2455C60.6005 29.2109 60.2164 30.1365 59.5337 30.8192C58.851 31.5018 57.9253 31.8857 56.9599 31.8867Z" fill="#79DAE8" />
+                                        <path d="M52.8194 39.8084V47.3437C52.8152 51.152 51.3006 54.8032 48.6077 57.4962C45.9149 60.1892 42.2638 61.7041 38.4554 61.7085C30.8432 61.7085 24.5133 55.6427 24.1209 48.0879C33.5968 47.0107 40.9468 38.8324 40.9471 29.2479V9.37539C40.9464 8.81397 40.7231 8.27573 40.3262 7.87875C39.9292 7.48176 39.3909 7.25847 38.8295 7.25784H34.1584C33.689 6.40902 32.9508 5.74026 32.0598 5.3567C31.1689 4.97314 30.1758 4.89654 29.2366 5.13892C28.2974 5.38131 27.4654 5.92895 26.8713 6.69574C26.2773 7.46253 25.9549 8.405 25.9548 9.37498C25.9548 10.345 26.2771 11.2875 26.8711 12.0543C27.4651 12.8211 28.2971 13.3688 29.2363 13.6113C30.1755 13.8537 31.1686 13.7772 32.0596 13.3937C32.9505 13.0102 33.6888 12.3414 34.1583 11.4927H36.7127V29.2479C36.7145 31.1847 36.3347 33.1028 35.5948 34.8927C34.8549 36.6826 33.7695 38.3091 32.4007 39.6792C31.0318 41.0494 29.4064 42.1364 27.6172 42.878C25.8281 43.6196 23.9103 44.0013 21.9735 44.0013C20.0368 44.0013 18.119 43.6196 16.3298 42.878C14.5407 42.1364 12.9152 41.0494 11.5464 39.6792C10.1775 38.3091 9.09216 36.6826 8.35228 34.8927C7.61241 33.1028 7.23255 31.1847 7.23441 29.2479V11.4927H9.78881C10.2583 12.3414 10.9966 13.0101 11.8875 13.3936C12.7785 13.7771 13.7716 13.8536 14.7108 13.6111C15.65 13.3686 16.482 12.8209 17.076 12.0541C17.6699 11.2872 17.9922 10.3447 17.9921 9.37472C17.9921 8.40474 17.6697 7.46228 17.0756 6.69551C16.4815 5.92873 15.6495 5.38112 14.7103 5.13877C13.7711 4.89641 12.7779 4.97305 11.887 5.35663C10.9961 5.74022 10.2579 6.409 9.78854 7.25784H5.1174C4.55601 7.25849 4.01779 7.4818 3.62084 7.87878C3.22389 8.27576 3.00061 8.81399 3 9.37539V29.2479C3.00014 38.8559 10.3773 47.0425 19.8825 48.0938C20.2812 57.9792 28.5145 65.9425 38.4554 65.9425C48.7104 65.9425 57.0538 57.5995 57.0538 47.3437V39.8084C55.6557 40.0829 54.2176 40.0829 52.8194 39.8084Z" fill="#79DAE8" />
+                                    </svg>
                                 </div>
-                            </div>
-                        </div>
-                        <!--/Enlace principal-->
-            
-                        <!--Enlace principal-->
-                        <div class="nav-link d-none d-lg-flex flex-row align-items-center">
-                            <div class="icon-estetoscopio mr-3">
-                                <svg width="70" height="70" viewBox="0 0 70 70" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                    <path fill-rule="evenodd" clip-rule="evenodd" d="M35 4.375C28.6038 4.375 23.4106 9.56812 23.4106 15.9644C23.4106 22.3606 28.6038 27.5537 35 27.5537C41.3963 27.5537 46.5894 22.3606 46.5894 15.9644C46.5894 9.56812 41.3963 4.375 35 4.375ZM35 7.65625C39.585 7.65625 43.3081 11.3794 43.3081 15.9644C43.3081 20.5494 39.585 24.2725 35 24.2725C30.415 24.2725 26.6919 20.5494 26.6919 15.9644C26.6919 11.3794 30.415 7.65625 35 7.65625Z" fill="#79DAE8" />
-                                    <path fill-rule="evenodd" clip-rule="evenodd" d="M16.9531 48.6828C16.9531 49.5884 17.6881 50.3234 18.5938 50.3234H51.4062C52.3119 50.3234 53.0469 49.5884 53.0469 48.6828V42.3128C53.0469 37.5265 51.1459 32.9371 47.7619 29.5553C44.3778 26.1712 39.7884 24.2703 35.0044 24.2703C35.0022 24.2703 34.9978 24.2703 34.9956 24.2703C30.2116 24.2703 25.6222 26.1712 22.2381 29.5553C18.8541 32.9371 16.9531 37.5265 16.9531 42.3128V48.6828ZM20.2344 47.0421V42.3128C20.2344 38.3971 21.7897 34.6434 24.5569 31.874C27.3263 29.1068 31.08 27.5515 34.9956 27.5515H35.0044C38.92 27.5515 42.6737 29.1068 45.4431 31.874C48.2103 34.6434 49.7656 38.3971 49.7656 42.3128V47.0421H20.2344Z" fill="#79DAE8" />
-                                    <path fill-rule="evenodd" clip-rule="evenodd" d="M67.2658 54.1407V53.0579C67.2658 49.735 64.573 47.0422 61.2501 47.0422C49.9801 47.0422 20.0552 47.0422 8.78515 47.0422C5.46234 47.0422 2.76953 49.735 2.76953 53.0579V54.1407C2.76953 57.4635 5.46234 60.1563 8.78515 60.1563H61.2501C64.573 60.1563 67.2658 57.4635 67.2658 54.1407ZM63.9845 53.0579V54.1407C63.9845 55.6501 62.7595 56.875 61.2501 56.875H8.78515C7.27578 56.875 6.05078 55.6501 6.05078 54.1407V53.0579C6.05078 51.5463 7.27578 50.3235 8.78515 50.3235H61.2501C62.7595 50.3235 63.9845 51.5463 63.9845 53.0579Z" fill="#79DAE8" />
-                                    <path fill-rule="evenodd" clip-rule="evenodd" d="M15.877 65.625H54.1582C55.0638 65.625 55.7988 64.89 55.7988 63.9844C55.7988 63.0788 55.0638 62.3438 54.1582 62.3438H15.877C14.9713 62.3438 14.2363 63.0788 14.2363 63.9844C14.2363 64.89 14.9713 65.625 15.877 65.625Z" fill="#79DAE8" />
-                                    <path fill-rule="evenodd" clip-rule="evenodd" d="M30.3339 32.7709C28.5424 32.7928 27.0942 34.2519 27.0942 36.05V38.71C27.0942 40.5081 28.5424 41.9694 30.3339 41.9912C30.3558 43.7828 31.8171 45.2309 33.6152 45.2309H36.2752C38.0733 45.2309 39.5324 43.7828 39.5542 41.9912C41.348 41.9694 42.7939 40.5081 42.7939 38.71V36.05C42.7939 34.2519 41.348 32.7928 39.5542 32.7709C39.5324 30.9772 38.0733 29.5312 36.2752 29.5312H33.6152C31.8171 29.5312 30.3558 30.9772 30.3339 32.7709ZM36.2752 36.05V32.8125H33.6152V36.05H30.3755V38.71H33.6152V41.9497H36.2752V38.71H39.5127V36.05H36.2752Z" fill="#79DAE8" />
-                                </svg>
-                            </div>
-            
-                            <div class="dropdown">
-                                <button class="btn_drop" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">Soy Paciente</button>
-                                <div class="dropdown-menu" aria-labelledby="dropdownMenuButton">
-                                    <button class="dropdown-item" data-toggle="modal" data-target="#loginPaciente">Ingresar</button>
-                                    <a class="dropdown-item" href="SignInPatient.html">Registrarse</a>
+                                <div class="">Soy Doctor</div>
+                            </a>
+                            <a href="SignInPatient.html" class="nav-link d-none d-lg-flex flex-row align-items-center">
+                                <div class="icon-estetoscopio mr-3">
+                                    <svg width="70" height="70" viewBox="0 0 70 70" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                        <path fill-rule="evenodd" clip-rule="evenodd" d="M35 4.375C28.6038 4.375 23.4106 9.56812 23.4106 15.9644C23.4106 22.3606 28.6038 27.5537 35 27.5537C41.3963 27.5537 46.5894 22.3606 46.5894 15.9644C46.5894 9.56812 41.3963 4.375 35 4.375ZM35 7.65625C39.585 7.65625 43.3081 11.3794 43.3081 15.9644C43.3081 20.5494 39.585 24.2725 35 24.2725C30.415 24.2725 26.6919 20.5494 26.6919 15.9644C26.6919 11.3794 30.415 7.65625 35 7.65625Z" fill="#79DAE8" />
+                                        <path fill-rule="evenodd" clip-rule="evenodd" d="M16.9531 48.6828C16.9531 49.5884 17.6881 50.3234 18.5938 50.3234H51.4062C52.3119 50.3234 53.0469 49.5884 53.0469 48.6828V42.3128C53.0469 37.5265 51.1459 32.9371 47.7619 29.5553C44.3778 26.1712 39.7884 24.2703 35.0044 24.2703C35.0022 24.2703 34.9978 24.2703 34.9956 24.2703C30.2116 24.2703 25.6222 26.1712 22.2381 29.5553C18.8541 32.9371 16.9531 37.5265 16.9531 42.3128V48.6828ZM20.2344 47.0421V42.3128C20.2344 38.3971 21.7897 34.6434 24.5569 31.874C27.3263 29.1068 31.08 27.5515 34.9956 27.5515H35.0044C38.92 27.5515 42.6737 29.1068 45.4431 31.874C48.2103 34.6434 49.7656 38.3971 49.7656 42.3128V47.0421H20.2344Z" fill="#79DAE8" />
+                                        <path fill-rule="evenodd" clip-rule="evenodd" d="M67.2658 54.1407V53.0579C67.2658 49.735 64.573 47.0422 61.2501 47.0422C49.9801 47.0422 20.0552 47.0422 8.78515 47.0422C5.46234 47.0422 2.76953 49.735 2.76953 53.0579V54.1407C2.76953 57.4635 5.46234 60.1563 8.78515 60.1563H61.2501C64.573 60.1563 67.2658 57.4635 67.2658 54.1407ZM63.9845 53.0579V54.1407C63.9845 55.6501 62.7595 56.875 61.2501 56.875H8.78515C7.27578 56.875 6.05078 55.6501 6.05078 54.1407V53.0579C6.05078 51.5463 7.27578 50.3235 8.78515 50.3235H61.2501C62.7595 50.3235 63.9845 51.5463 63.9845 53.0579Z" fill="#79DAE8" />
+                                        <path fill-rule="evenodd" clip-rule="evenodd" d="M15.877 65.625H54.1582C55.0638 65.625 55.7988 64.89 55.7988 63.9844C55.7988 63.0788 55.0638 62.3438 54.1582 62.3438H15.877C14.9713 62.3438 14.2363 63.0788 14.2363 63.9844C14.2363 64.89 14.9713 65.625 15.877 65.625Z" fill="#79DAE8" />
+                                        <path fill-rule="evenodd" clip-rule="evenodd" d="M30.3339 32.7709C28.5424 32.7928 27.0942 34.2519 27.0942 36.05V38.71C27.0942 40.5081 28.5424 41.9694 30.3339 41.9912C30.3558 43.7828 31.8171 45.2309 33.6152 45.2309H36.2752C38.0733 45.2309 39.5324 43.7828 39.5542 41.9912C41.348 41.9694 42.7939 40.5081 42.7939 38.71V36.05C42.7939 34.2519 41.348 32.7928 39.5542 32.7709C39.5324 30.9772 38.0733 29.5312 36.2752 29.5312H33.6152C31.8171 29.5312 30.3558 30.9772 30.3339 32.7709ZM36.2752 36.05V32.8125H33.6152V36.05H30.3755V38.71H33.6152V41.9497H36.2752V38.71H39.5127V36.05H36.2752Z" fill="#79DAE8" />
+                                    </svg>
                                 </div>
+                                <div>Soy Paciente</div>
+                            </a>
+                            
+                            <div class="ml-lg-5">
+                                <button id="main_hamburger_menu_btn">
+                                    <svg width="70" height="70" viewBox="0 0 70 70" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                        <rect width="70" height="70" rx="7" fill="#79DAE8" />
+                                        <path d="M57.9688 22.9688H12.0312C10.1719 22.9688 8.75 21.5469 8.75 19.6875C8.75 17.8281 10.1719 16.4062 12.0312 16.4062H57.9688C59.8281 16.4062 61.25 17.8281 61.25 19.6875C61.25 21.5469 59.8281 22.9688 57.9688 22.9688Z" fill="white" />
+                                        <path d="M57.9688 38.2812H12.0312C10.1719 38.2812 8.75 36.8594 8.75 35C8.75 33.1406 10.1719 31.7188 12.0312 31.7188H57.9688C59.8281 31.7188 61.25 33.1406 61.25 35C61.25 36.8594 59.8281 38.2812 57.9688 38.2812Z" fill="white" />
+                                        <path d="M57.9688 53.5938H12.0312C10.1719 53.5938 8.75 52.1719 8.75 50.3125C8.75 48.4531 10.1719 47.0312 12.0312 47.0312H57.9688C59.8281 47.0312 61.25 48.4531 61.25 50.3125C61.25 52.1719 59.8281 53.5938 57.9688 53.5938Z" fill="white" />
+                                    </svg>
+                                </button>
                             </div>
-                        </div>
-                        <!--Enlace principal-->
-            
-                        <div class="ml-lg-5">
-                            <button id="main_hamburger_menu_btn">
-                                <svg width="70" height="70" viewBox="0 0 70 70" fill="none" xmlns="http://www.w3.org/2000/svg">
-                                    <rect width="70" height="70" rx="7" fill="#79DAE8" />
-                                    <path d="M57.9688 22.9688H12.0312C10.1719 22.9688 8.75 21.5469 8.75 19.6875C8.75 17.8281 10.1719 16.4062 12.0312 16.4062H57.9688C59.8281 16.4062 61.25 17.8281 61.25 19.6875C61.25 21.5469 59.8281 22.9688 57.9688 22.9688Z" fill="white" />
-                                    <path d="M57.9688 38.2812H12.0312C10.1719 38.2812 8.75 36.8594 8.75 35C8.75 33.1406 10.1719 31.7188 12.0312 31.7188H57.9688C59.8281 31.7188 61.25 33.1406 61.25 35C61.25 36.8594 59.8281 38.2812 57.9688 38.2812Z" fill="white" />
-                                    <path d="M57.9688 53.5938H12.0312C10.1719 53.5938 8.75 52.1719 8.75 50.3125C8.75 48.4531 10.1719 47.0312 12.0312 47.0312H57.9688C59.8281 47.0312 61.25 48.4531 61.25 50.3125C61.25 52.1719 59.8281 53.5938 57.9688 53.5938Z" fill="white" />
-                                </svg>
-                            </button>
                         </div>
                     </div>
                 </div>
-            </div>
-            
             </nav>
         </div>
     `;
